types: webrtc: Rename file for consistency composables: Update to use new types/webRTC
